refactor colorProvider + update highlighting data
